+++ src/AssetTrash.js
@@ -9,7 +9,7 @@ class AssetTrash {
 
   /**
    * Reset settings.
-   * This method is called before each compilation after changes by `webpack serv/watch`.
+   * Called before each compilation after changes by `webpack serv/watch`.
    */
   static reset() {
     this.trash = [];